+++ src/test/java/racinggame/domain/CarTest.java
@@ -1,10 +1,10 @@
-package racinggame;
+package racinggame.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racinggame.domain.Car;
 
 public class CarTest {
