+++ src/test/java/racinggame/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
+import racinggame.domain.Car;
+import racinggame.domain.Cars;
 
 public class CarsTest {
