+++ src/test/java/racinggame/domain/SnapShotStoreTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,8 +6,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 import org.junit.jupiter.api.Test;
-import racinggame.Car.CarDto;
-import racinggame.SnapShotStore.SnapShot;
+import racinggame.domain.Car.CarDto;
+import racinggame.domain.SnapShotStore;
+import racinggame.domain.SnapShotStore.SnapShot;
 
 public class SnapShotStoreTest {
